Добавлена реализация кэширования сущностей (Request) через Redis. Добавлена возможность соединения с БД в несколько попыток.
