+++ Dockerfile.api
@@ -15,9 +15,9 @@ LABEL org.opencontainers.image.source https://${GITHUB_PATH}
 RUN apk --no-cache add ca-certificates
 WORKDIR /root/
 
-COPY --from=builder /home/${GITHUB_PATH}/bin/grpc-server .
-COPY --from=builder /home/${GITHUB_PATH}/config.yml .
-COPY --from=builder /home/${GITHUB_PATH}/migrations/migrations ./migrations
+COPY --from=builder /home/${GITHUB_PATH}/bin/grpc-server ./
+COPY --from=builder /home/${GITHUB_PATH}/config.yml ./
+COPY --from=builder /home/${GITHUB_PATH}/migrations/migrations ./migrations/
 
 RUN chown root:root grpc-server
