+++ cmd/retranslator/main.go
@@ -53,7 +53,7 @@ func main() {
 		cfg.Database.SslMode,
 	)
 
-	db, err := database.NewPostgres(dsn, cfg.Database.Driver)
+	db, err := database.NewPostgres(dsn, cfg.Database.Driver, cfg.Database.Retry)
 	if err != nil {
 		log.Fatal().Err(err).Msg("Failed init postgres")
 	}
