+++ docker-compose.yaml
@@ -15,8 +15,10 @@ services:
       - ompnw
     depends_on:
       - postgres
+      - redis
     links:
       - postgres
+      - redis
     ports:
       - 8080:8080 # REST
       - 8082:8082 # gRPC
@@ -57,6 +59,17 @@ services:
       - kafka-3
     volumes:
       - ./config.yml:/root/config.yml
+
+  redis:
+    image: redis:latest
+    ports:
+      - 6379:6379
+    volumes:
+      - ../redis-data:/var/lib/redis
+    environment:
+      - REDIS_REPLICATION_MODE=master
+    networks:
+      - ompnw
 
   postgres:
     image: postgres:latest
